LayoutTests/platform/chromium-mac/fast/encoding/char-encoding-mac-expected.txt

This tests encoding characters in various character sets.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS encode('mac', 'U+221A') is '%C3'
PASS encode('macintosh', 'U+221A') is '%C3'
PASS encode('MacRoman', 'U+221A') is '%C3'
FAIL encode('x-mac-greek', 'U+0391') should be %B0. Was %26%23913%3B.
PASS encode('x-mac-cyrillic', 'U+0410') is '%80'
PASS encode('mac-cyrillic', 'U+0410') is '%80'
FAIL encode('x-mac-centraleurroman', 'U+0100') should be %81. Was %26%23256%3B.
FAIL encode('x-mac-turkish', 'U+00C5') should be %81. Was %C5.
PASS successfullyParsed is true

TEST COMPLETE
